// apps/frontend/src/api/generated/model/incomeSourceSubType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.37
 */

export type IncomeSourceSubType = (typeof IncomeSourceSubType)[keyof typeof IncomeSourceSubType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const IncomeSourceSubType = {
  payroll: 'payroll',
  bonus: 'bonus',
  rsu: 'rsu',
  reimbursement: 'reimbursement',
} as const