// apps/frontend/src/api/generated/model/variabilityModel.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.6
 */

export type VariabilityModel = (typeof VariabilityModel)[keyof typeof VariabilityModel]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VariabilityModel = {
  fixed: 'fixed',
  range: 'range',
  historical_distribution: 'historical_distribution',
} as const